writing the README. Minor changes in project structure.
